refactor(sql): parse constraints to domain objects, colocate type mapping, rename describe module
Three follow-up cleanups to the AS JSON read path:

- constraints.py now builds the domain PrimaryKeyConstraint / ForeignKeyConstraint
  directly. Deletes the shallow ParsedPrimaryKey / ParsedForeignKey mirrors and
  describe's _lower_constraints re-box. A loop-level ValueError -> ConstraintParseError
  keeps the parser's "malformed -> typed error" contract when a domain constructor
  rejects a key. ParsedConstraints stays as a thin named result holder.

- Move data_type_from_json (structured JSON -> domain) into types.py, beside
  render_data_type (domain -> DDL), so one module owns both directions of the
  type mapping. Also fixes types.py's docstring, which still pointed at the
  deleted parse.py as the read half.

- Rename describe_json.py -> describe.py (and its test); it now handles document
  structure only and delegates type mapping to types.py. describe_json_query in
  queries.py keeps its name (it names the literal DESCRIBE ... AS JSON SQL).
