+++ tests/adapters/databricks/sql/test_types.py
@@ -1,4 +1,4 @@
-from delta_engine.adapters.databricks.sql.types import render_data_type
+from delta_engine.adapters.databricks.sql.types import data_type_from_json, render_data_type
 from delta_engine.domain.model.data_type import (
     Array,
     Binary,
@@ -49,3 +49,80 @@ def test_sql_type_for_decimal_array_map_recursive() -> None:
 def test_sql_type_for_struct_renders_fields_in_order() -> None:
     struct = Struct((StructField("a", Integer()), StructField("b", Array(String()))))
     assert render_data_type(struct) == "STRUCT<`a`: INT, `b`: ARRAY<STRING>>"
+
+
+def test_primitive_aliases():
+    assert data_type_from_json({"name": "int"}) == Integer()
+    assert data_type_from_json({"name": "integer"}) == Integer()
+    assert data_type_from_json({"name": "bigint"}) == Long()
+    assert data_type_from_json({"name": "double"}) == Double()
+    assert data_type_from_json({"name": "boolean"}) == Boolean()
+
+
+def test_string_ignores_collation_and_length():
+    assert data_type_from_json({"name": "string", "collation": "UTF8_BINARY"}) == String()
+    assert data_type_from_json({"name": "varchar", "length": 20}) == String()
+
+
+def test_timestamp_ltz_aliases_to_timestamp():
+    assert data_type_from_json({"name": "timestamp"}) == Timestamp()
+    assert data_type_from_json({"name": "timestamp_ltz"}) == Timestamp()
+    assert data_type_from_json({"name": "timestamp_ntz"}) == TimestampNtz()
+
+
+def test_decimal_reads_precision_and_scale():
+    assert data_type_from_json({"name": "decimal", "precision": 10, "scale": 2}) == Decimal(10, 2)
+
+
+def test_array_map_struct_nested():
+    assert data_type_from_json(
+        {"name": "array", "element_type": {"name": "string"}, "element_nullable": True}
+    ) == Array(String())
+    assert data_type_from_json(
+        {"name": "map", "key_type": {"name": "string"}, "value_type": {"name": "int"}}
+    ) == Map(String(), Integer())
+    assert data_type_from_json(
+        {
+            "name": "struct",
+            "fields": [
+                {"name": "Age", "type": {"name": "int"}, "nullable": True},
+                {"name": "label", "type": {"name": "string"}, "nullable": True},
+            ],
+        }
+    ) == Struct((StructField("age", Integer()), StructField("label", String())))
+
+
+def test_unmappable_returns_none():
+    assert data_type_from_json({"name": "interval"}) is None
+    assert (
+        data_type_from_json(
+            {
+                "name": "struct",
+                "fields": [
+                    {"name": "a", "type": {"name": "int"}},
+                    {"name": "A", "type": {"name": "int"}},
+                ],
+            }
+        )
+        is None
+    )  # duplicate field name after casefold
+    assert data_type_from_json({"not": "a type"}) is None
+    assert data_type_from_json("string") is None
+
+
+def test_blank_struct_field_name_returns_none():
+    assert (
+        data_type_from_json({"name": "struct", "fields": [{"name": "  ", "type": {"name": "int"}}]})
+        is None
+    )
+
+
+def test_decimal_over_delta_limit_returns_none():
+    assert data_type_from_json({"name": "decimal", "precision": 40, "scale": 2}) is None
+
+
+def test_pathologically_deep_nesting_returns_none():
+    payload = {"name": "int"}
+    for _ in range(6000):
+        payload = {"name": "array", "element_type": payload}
+    assert data_type_from_json(payload) is None
